move modal component to top level
