+++ deepfence_frontend/apps/dashboard/src/features/registries/components/RegistryAccountsTable.tsx
@@ -28,6 +28,7 @@ import {
 
 import { getRegistriesApiClient } from '@/api/api';
 import { ApiDocsBadRequestResponse, ModelRegistryListResp } from '@/api/generated';
+import { ConfigureScanModal } from '@/components/ConfigureScanModal';
 import { DFLink } from '@/components/DFLink';
 import { MalwareScanActionEnumType } from '@/components/scan-configure-forms/MalwareScanConfigureForm';
 import { SecretScanActionEnumType } from '@/components/scan-configure-forms/SecretScanConfigureForm';
@@ -35,7 +36,6 @@ import { VulnerabilityScanActionEnumType } from '@/components/scan-configure-for
 import { MalwareIcon } from '@/components/sideNavigation/icons/Malware';
 import { SecretsIcon } from '@/components/sideNavigation/icons/Secrets';
 import { VulnerabilityIcon } from '@/components/sideNavigation/icons/Vulnerability';
-import { ConfigureScanModal } from '@/features/registries/components/ConfigureScanModal';
 import { ApiError, makeRequest } from '@/utils/api';
 import { formatMilliseconds } from '@/utils/date';
