+++ deepfence_frontend/apps/dashboard/src/features/registries/components/RegistryImageTagsTable.tsx
@@ -17,13 +17,13 @@ import {
 } from 'ui-components';
 
 import { ModelContainerImage } from '@/api/generated';
+import { ConfigureScanModal } from '@/components/ConfigureScanModal';
 import { MalwareScanActionEnumType } from '@/components/scan-configure-forms/MalwareScanConfigureForm';
 import { SecretScanActionEnumType } from '@/components/scan-configure-forms/SecretScanConfigureForm';
 import { VulnerabilityScanActionEnumType } from '@/components/scan-configure-forms/VulnerabilityScanConfigureForm';
 import { MalwareIcon } from '@/components/sideNavigation/icons/Malware';
 import { SecretsIcon } from '@/components/sideNavigation/icons/Secrets';
 import { VulnerabilityIcon } from '@/components/sideNavigation/icons/Vulnerability';
-import { ConfigureScanModal } from '@/features/registries/components/ConfigureScanModal';
 import { formatMilliseconds } from '@/utils/date';
 
 const PAGE_SIZE = 15;
